+++ build.clj
@@ -84,7 +84,7 @@
                         "--no-fallback"
                         "--native-image-info"
                         "--features=clj_easy.graal_build_time.InitClojureClasses"
-                        (when-not (fs/windows?) "-march=compatibility")
+                        "-march=compatibility"
                         "-O1"
                         (or (System/getenv "ECA_XMX")
                             "-J-Xmx8g")