+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix native image on Windows not running on older CPUs by enabling `-march=compatibility`. #362
+- Fix simple summaries in tool calls for ollama.
 
 ## 0.114.0